requirements.txt
supabase==2.10.0
speedtest-cli==2.1.3
python-dotenv==1.0.0